\name{ISgeometry}
\alias{ISgeometry}

\title{ISgeometry}

\description{Incoherent scatter radar measurement geometry evaluation}

\details{Estimates of incoherent integration times, assuming low SNR and
  Gaussian beam shpes

  Following coordinates are loaded to the global workspace at the time
  when the package is loaded, and are available for the function calls:

  Tromso       TRO  = c(69.58,19.23)
  
  Kiruna       KIR  = c(67.87,20.43)

  Sodankyla    SOD  = c(67.37,26.63)

  Andoya       AND  = c(69.13,15.86)

  Esrange      ESRA = c(67.88,21.12)

  Lokkavaara   LOK  = c(67.81,27.69)

  Muonio       MUO  = c(67.96,23.68)

  Kilpisjarvi  KIL  = c(69.02,20.86)

  Säytsjärvi   SAY  = c(69.35,27.20)

  Suolojavrras SUO  = c(69.57,23.57)

  Miellejokk   MIE  = c(68.34,18.96)

  Parivierra   PAR  = c(67.06,19.59)

  Järämä       JAR  = c(68.38,21.08)

  Kautokeino   KAU  = c(69.01,23.04)

  Overbygd     OVE  = c(69.03,19.29)

  Abisko 	   ABI = c(68.37,18.75)

  Skibotn      SKI = c(69.58,20.47)

  Longyearbyen ESR = c(78.15,16.03)

  Sekkujärvi   SKJ = c(68.11,21.00)

  Järämä       JRM = c(68.39,21.01)

  Karasjokk    KJK = c(69.46,25.53)

  Alta         ALT = c(69.86,22.96)

  Masi         MAS = c(69.44,23.65)

  Altaelva     AEV = c(69.19,23.57)

  Hetta        HTA = c(68.39,23.60)

  Palojoensuu  PLJ = c(68.28,23.08)

  Karesuvanto  KRS = c(68.45,22.48)

  Peera        PEE = c(68.88,21.06)



  FUNCTIONS:
  
  \code{\link{multistaticIntegrationTimes}} integration time evaluation
  
  \code{\link{integrationTimeVsArea}} calculates and/or plots the covered area as function of integration time

  \code{\link{savepdf}} save contents of a figure window in a pdf file

}



\author{Ilkka Virtanen (University of Oulu, Finland) \email{ilkka.i.virtanen@oulu.fi}}


