\name{savepdf}
\alias{savepdf}
\title{savepdf}


\description{Save contents of a figure window into a pdf file}

\usage{
savepdf(devnum, fname)
}

\arguments{
\item{devnum}{The graphics device whose contents will be copied to file,
  e.g. the number on the top of each graphics window}
\item{fname}{File name to write. The function does not add the file name
extension, so remember to include '.pdf' in fname.}

  }

  \value{None}

  \details{}

  \author{Ilkka Virtanen \email{ilkka.i.virtanen@oulu.fi}}